[fix]차트 파일명 변경
